move things inside blocklist module
