+++ packages/twenty-server/src/modules/blocklist/blocklist-validation.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
 import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
-import { BlocklistValidationService } from 'src/modules/connected-account/services/blocklist/blocklist-validation.service';
-import { BlocklistWorkspaceEntity } from 'src/modules/connected-account/standard-objects/blocklist.workspace-entity';
+import { BlocklistValidationService } from 'src/modules/blocklist/services/blocklist-validation.service';
+import { BlocklistWorkspaceEntity } from 'src/modules/blocklist/standard-objects/blocklist.workspace-entity';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
 
 @Module({
