+++ packages/twenty-server/src/modules/blocklist/query-hooks/blocklist-query-hook.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+
+import { BlocklistValidationModule } from 'src/modules/blocklist/blocklist-validation.module';
+import { BlocklistCreateManyPreQueryHook } from 'src/modules/blocklist/query-hooks/blocklist-create-many.pre-query.hook';
+import { BlocklistUpdateManyPreQueryHook } from 'src/modules/blocklist/query-hooks/blocklist-update-many.pre-query.hook';
+import { BlocklistUpdateOnePreQueryHook } from 'src/modules/blocklist/query-hooks/blocklist-update-one.pre-query.hook';
+
+@Module({
+  imports: [BlocklistValidationModule],
+  providers: [
+    BlocklistCreateManyPreQueryHook,
+    BlocklistUpdateManyPreQueryHook,
+    BlocklistUpdateOnePreQueryHook,
+  ],
+})
+export class BlocklistQueryHookModule {}
