+++ packages/twenty-server/src/modules/blocklist/query-hooks/blocklist-update-many.pre-query.hook.ts
