+++ packages/twenty-server/src/modules/blocklist/query-hooks/blocklist-update-one.pre-query.hook.ts
@@ -1,14 +1,14 @@
 import { WorkspaceQueryHookInstance } from 'src/engine/api/graphql/workspace-query-runner/workspace-query-hook/interfaces/workspace-query-hook.interface';
-import { CreateManyResolverArgs } from 'src/engine/api/graphql/workspace-resolver-builder/interfaces/workspace-resolvers-builder.interface';
+import { UpdateOneResolverArgs } from 'src/engine/api/graphql/workspace-resolver-builder/interfaces/workspace-resolvers-builder.interface';
 
 import { WorkspaceQueryHook } from 'src/engine/api/graphql/workspace-query-runner/workspace-query-hook/decorators/workspace-query-hook.decorator';
 import {
   BlocklistItem,
   BlocklistValidationService,
-} from 'src/modules/connected-account/services/blocklist/blocklist-validation.service';
+} from 'src/modules/blocklist/services/blocklist-validation.service';
 
-@WorkspaceQueryHook(`blocklist.createMany`)
-export class BlocklistCreateManyPreQueryHook
+@WorkspaceQueryHook(`blocklist.updateOne`)
+export class BlocklistUpdateOnePreQueryHook
   implements WorkspaceQueryHookInstance
 {
   constructor(
@@ -18,9 +18,9 @@ export class BlocklistCreateManyPreQueryHook
   async execute(
     userId: string,
     workspaceId: string,
-    payload: CreateManyResolverArgs<BlocklistItem>,
+    payload: UpdateOneResolverArgs<BlocklistItem>,
   ): Promise<void> {
-    await this.blocklistValidationService.validateBlocklistForCreateMany(
+    await this.blocklistValidationService.validateBlocklistForUpdateOne(
       payload,
       userId,
       workspaceId,
