+++ packages/twenty-server/src/modules/blocklist/standard-objects/blocklist.workspace-entity.ts
